requirements.txt
streamlit==1.38.0
openai==1.51.2
feedparser==6.0.11
requests==2.32.3
python-dotenv==1.0.1
gspread==6.1.2
google-auth==2.35.0